// src/models/User.ts
export interface User {
  id?: string
  email?: string
  firstName?: string
  lastName?: string
  fullName?: string
  nickname?: string
  role?: string
  avatarUrl?: string
  profileUrl?: string
  gender?: boolean
  resigned?: boolean
  address?: string
  phone?: string
  position?: string
  dateOfBirth?: string
}